+++ playwright/pageobjects/personal.page.ts
@@ -9,7 +9,6 @@ export class PersonalPage {
   readonly checkBox: Locator;
   readonly previousButton: Locator;
   readonly nextButton: Locator;
-
   readonly firstNameErrorMessage: Locator;
   readonly lastNameErrorMessage: Locator;
   readonly emailErrorMessage: Locator;
@@ -26,10 +25,7 @@ export class PersonalPage {
         "Yes, I would like to receive information about NYPL's programs and services",
     });
     this.previousButton = this.page.getByRole("link", { name: "Previous" });
-    this.nextButton = this.page.getByRole("button", {
-      name: "Next",
-      exact: true,
-    });
+    this.nextButton = this.page.getByRole("button", { name: "Next" });
 
     this.firstNameErrorMessage = this.page.getByText(
       "Please enter a valid first name."